+++ wallet-service/src/main/java/com/wlt/wallet/dto/TransactionHistoryEvent.java
@@ -21,7 +21,7 @@ public class TransactionHistoryEvent {
     private BigDecimal amount;
     private BigDecimal crAmount;
     private BigDecimal drAmount;
-    private BigDecimal exchangeRate;
+    private float exchangeRate;
     private LocalDateTime saveTimestamp;
     private String status;
 }
